refactor: extract the Slack reporter into internal/slack
Fifth leaf of the package split: SlackReporter and SlackFinding move
to internal/slack as Reporter and Finding, along with the 38
reporter-side tests (formatting, webhook posting, dedup, state
persistence, migration). The agent-side Check* tests stay in
pkg/agent, which bridges the old names with aliases.

Two shared helpers surfaced by the move find their proper homes: the
GitHub web-URL builders (PRURL, CommitURL, IssueURL) land in
internal/gh with their test, used by both the reporter and the
agent's finding collectors; shortSHA gets a local copy for the report
header. plugin_test.go no longer pokes the reporter's unexported
httpClient - the default client already reaches the plain-HTTP test
server.
